moved classes in better named packages
