+++ infra/src/edu/illinois/gitsvn/analysis/launchers/PureReposAnalysis.java
@@ -1,10 +1,10 @@
-package edu.illinois.gitsvn.analysis;
+package edu.illinois.gitsvn.analysis.launchers;
 
 import edu.illinois.gitsvn.infra.DirectoryAnalysisLauncher;
 
 public class PureReposAnalysis {
 	public static void main(String[] args) throws Exception {
 		new DirectoryAnalysisLauncher("../../pureGitRepos").run();
-		//new DirectoryAnalysisLauncher("../../pureSVNRepos");
+		// new DirectoryAnalysisLauncher("../../pureSVNRepos");
 	}
 }
